move base loaders to spatial package
