+++ packages/spatial/src/resources/AssetType.test.ts
