+++ packages/spatial/src/resources/AssetType.ts
