+++ packages/spatial/src/resources/DomainConfigState.ts
