+++ packages/spatial/src/resources/ResourceProgressComponent.ts
