+++ packages/spatial/src/resources/fileTypes.tsx
