+++ packages/spatial/src/resources/loaders/base/FileLoader.test.ts
