+++ packages/spatial/src/resources/loaders/base/Loader.ts
