+++ packages/spatial/src/resources/loaders/base/ResourceCache.ts
