+++ packages/spatial/src/resources/loaders/image/ImageBitmapLoader.ts
