+++ packages/spatial/src/resources/loaders/ktx2/KTX2Loader.d.ts
