+++ packages/spatial/src/resources/loaders/ktx2/KTX2LoaderState.ts
@@ -23,14 +23,32 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-export const DEFAULT_LOD_DISTANCES = Object.freeze({
-  '0': 5,
-  '1': 15,
-  '2': 30
-})
+import { defineState, getState, isClient } from '@ir-engine/hyperflux'
 
-export const LOADER_STATUS = {
-  LOADED: 0,
-  LOADING: 1,
-  ERROR: 2
-}
+import { WebGLRenderer } from 'three'
+import { DomainConfigState } from '../../DomainConfigState'
+import { KTX2Loader } from './KTX2Loader'
+
+export const KTX2LoaderState = defineState({
+  name: 'KTX2LoaderState',
+  initial: () => {
+    const ktx2Loader = new KTX2Loader()
+    ktx2Loader.setTranscoderPath(getState(DomainConfigState).publicDomain + '/loader_decoders/basis/')
+    if (isClient) {
+      const renderer = new WebGLRenderer()
+      ktx2Loader.detectSupport(renderer)
+      renderer.dispose()
+    } else {
+      // @ts-ignore - make nodejs happy
+      ktx2Loader.workerConfig = {
+        astcSupported: false,
+        etc1Supported: false,
+        etc2Supported: false,
+        dxtSupported: false,
+        bptcSupported: false,
+        pvrtcSupported: false
+      }
+    }
+    return ktx2Loader
+  }
+})
