+++ packages/spatial/src/resources/loaders/ktx2/ktx-parse.module.js
