+++ packages/spatial/src/resources/loaders/ktx2/zstddec.module.js
