+++ packages/spatial/src/resources/loaders/texture/TexturePatch.ts
