+++ packages/spatial/src/resources/loaders/tga/TGALoader.ts
