+++ packages/spatial/src/resources/resourceLoaderFunctions.test.tsx
