+++ packages/spatial/src/resources/resourceLoaderHooks.test.tsx
